Add many new features, including an ability to parse Cookie Clicker saves
It is now possible to pass a save file in to the program, and it can parse it
into it's represented object form, so avoid managing a secondary form of state.
Certain other features are also not available, such as named upgrades, certain minigame buffs,
and advanced predictive configuration options.